backend/requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
mysqlclient>=2.2
stripe>=8.0
Pillow>=10.0
openpyxl>=3.1
reportlab>=4.0
drf-spectacular>=0.27
PyYAML>=6.0
